+++ src/v3.0.0/RideWindow/RideWindow.ts
@@ -15,13 +15,13 @@ let updateViewport: (IDisposable | null) = null;
 
 export function showWindowRide(ride: Ride) {
 	ride1 = ride;
-    console.log()
     viewport = getAllVehiclesAndStations(ride1);
 	if (windowShowRide) {
 		windowShowRide.bringToFront();
 		return;
 	}
-	const windowDesc: WindowDesc = {
+
+    const windowDesc: WindowDesc = {
 		classification: windowTag,
 		width: 475,
 		height: 720,
@@ -101,6 +101,96 @@ export function showWindowRide(ride: Ride) {
                 width: 550,
                 height: 20,
                 text: "Downtime: " + ride1.satisfaction
+            },
+            {
+                name: 'MaxSpeed',
+                type: 'label',
+                x: 5,
+                y: 280,
+                width: 550,
+                height: 20,
+                text: "Max speed in MPH: " + ride1.maxSpeed
+            },
+            {
+                name: 'AVGSpeed',
+                type: 'label',
+                x: 5,
+                y: 300,
+                width: 550,
+                height: 20,
+                text: "Avg speed in MPH: " + ride1.averageSpeed
+            },
+            {
+                name: 'RideTime',
+                type: 'label',
+                x: 5,
+                y: 320,
+                width: 550,
+                height: 20,
+                text: "Ride time: " + ride1.rideTime
+            },
+            {
+                name: 'RideLength',
+                type: 'label',
+                x: 5,
+                y: 340,
+                width: 550,
+                height: 20,
+                text: "Ride length: " + ride1.rideLength
+            },
+            {
+                name: 'maxPositiveVerticalGs',
+                type: 'label',
+                x: 5,
+                y: 360,
+                width: 550,
+                height: 20,
+                text: "PvGs: " + ride1.maxPositiveVerticalGs
+            },
+            {
+                name: 'maxNegativeVerticalGs',
+                type: 'label',
+                x: 5,
+                y: 380,
+                width: 550,
+                height: 20,
+                text: "NvGs: " + ride1.maxNegativeVerticalGs
+            },
+            {
+                name: 'maxLateralGs',
+                type: 'label',
+                x: 5,
+                y: 400,
+                width: 550,
+                height: 20,
+                text: "Max lateral Gs: " + ride1.maxLateralGs
+            },
+            {
+                name: 'numDrops',
+                type: 'label',
+                x: 5,
+                y: 420,
+                width: 550,
+                height: 20,
+                text: "Number of drops: " + ride1.numDrops
+            },
+            {
+                name: 'highestDropHeight',
+                type: 'label',
+                x: 5,
+                y: 440,
+                width: 550,
+                height: 20,
+                text: "Highest drop height: " + ride1.highestDropHeight
+            },
+            {
+                name: 'totalAirTime',
+                type: 'label',
+                x: 5,
+                y: 460,
+                width: 550,
+                height: 20,
+                text: "Total air time: " + ride1.totalAirTime
             },
 			{
 				type: 'label',
@@ -198,7 +288,7 @@ export function showWindowRide(ride: Ride) {
 			{
 				name: 'viewportShowRide',
 				type: 'viewport',
-				x: 120,
+				x: 140,
 				y: 20,
 				width: 250,
 				height: 400,
@@ -208,7 +298,7 @@ export function showWindowRide(ride: Ride) {
 				name: 'dropdownShowRide',
 				type: 'dropdown',
                 selectedIndex: 0,
-				x: 120,
+				x: 140,
 				y: 420,
 				width: 250,
 				height: 30,
@@ -519,7 +609,27 @@ export function showWindowRide(ride: Ride) {
                 onChange: (index) => {
                     ride1.stationStyle = index;
                 },
-            }
+            },
+            {
+                name: 'LifthillSpeedLabel',
+                type: 'label',
+                x: 5,
+                y: 620,
+                width: 550,
+                height: 20,
+                text: "Lift hill speed in MPH: "
+            },
+            {
+                type: 'spinner',
+                name: 'spinnerLiftHillSpeed',
+                x: 140,
+                y: 620,
+                width: 100,
+                height: 20,
+                text: ride1.liftHillSpeed.toString(),
+                onIncrement: () => windowShowRide.findWidget<SpinnerWidget>('spinnerLiftHillSpeed').text = setLiftHillSpeed(ride1, "IN"),
+                onDecrement: () => windowShowRide.findWidget<SpinnerWidget>('spinnerLiftHillSpeed').text = setLiftHillSpeed(ride1, "DE")
+            },
 		],
 		onClose() {
             if (updateViewport != null) {
@@ -609,16 +719,18 @@ function getAllVehiclesAndStations(ride1: Ride) {
     }
     id = 1;
     for (const element of ride1.vehicles) {
-        let vehicle: CoordsXYZ = {
-            x: map.getEntity(element).x,
-            y: map.getEntity(element).y,
-            z: map.getEntity(element).z
-        }
         if (element != null) {
-            let viewportObject: ViewportArray = new ViewportArray(vehicle, "Vehicle " + id);
-            viewportObject.setVehicleNumber(element);
-            viewport.push(viewportObject);
-            id++;
+            if (map.getEntity(element) != null) {
+                let vehicle: CoordsXYZ = {
+                    x: map.getEntity(element).x,
+                    y: map.getEntity(element).y,
+                    z: map.getEntity(element).z
+                }
+                let viewportObject: ViewportArray = new ViewportArray(vehicle, "Vehicle " + id);
+                viewportObject.setVehicleNumber(element);
+                viewport.push(viewportObject);
+                id++;
+            }
         }
     }
     return viewport;
@@ -652,4 +764,19 @@ function followTrain(viewportNumber: number) {
 function disableUpdateViewport(){
     updateViewport?.dispose();
     updateViewport = null;
+}
+
+
+function setLiftHillSpeed(ride1: Ride, in1: string) : string {
+
+    let speed: number = ride1.liftHillSpeed;
+
+    if (in1 === 'IN' && speed >= ride1.minLiftHillSpeed && speed < ride1.maxLiftHillSpeed) {
+        speed++;
+    }
+    else if (in1 === 'DE' && speed > ride1.minLiftHillSpeed && speed <= ride1.maxLiftHillSpeed) {
+        speed--;
+    }
+    context.executeAction("ridesetsetting", { ride: ride1.id, setting: 8, value: speed, flags: 0 });
+    return speed.toString();
 }